backend/requirements.txt
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
transformers==4.35.0
torch>=2.0.0
newspaper3k==0.2.8
requests==2.31.0
python-dotenv==1.0.0 